backend/requirements.txt
asgiref==3.11.0
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
colorama==0.4.6
Django==5.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
qrcode==8.2
reportlab==4.4.10
requests==2.33.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.7.0
